fix(react): restore textarea autosize browser coverage
